+++ airone/spectacular.py
@@ -2,6 +2,7 @@
 # OpenAPI schema using drf-spectacular.
 # (c.f. https://drf-spectacular.readthedocs.io/en/latest/settings.html)
 
+
 def exclude_customview_hook(endpoints):
     """This excludes API endpoints of custom-view's ones. Because it's not always
     necessary to generage custom-view's OpenAPI schema.